21.03.2020 - DB of clients and cretid cards is generated with FK and good data, few methods in DAO are implemented

+++ src/main/java/services/entities/Payment.java
@@ -2,7 +2,7 @@ package services.entities;
 
 public class Payment {
     public int paymentId;
-    public int cardNumber;
+    public long cardNumber;
     public PaymentType paymentType;
     public int amount;
     public int cardBalance;
